+++ src/main/java/com/example/cinema/vo/RefundVO.java
@@ -17,12 +17,12 @@ public class RefundVO {
     /**
      * 适用时间，电影开场前多久
      */
-    private int time;
+    private double time;
 
     /**
      * 折算策略
      */
-    private int price;
+    private double price;
 
     public int getId() {
         return id;
@@ -40,19 +40,19 @@ public class RefundVO {
         this.movie = movie;
     }
 
-    public int getTime() {
+    public double getTime() {
         return time;
     }
 
-    public void setTime(int time) {
+    public void setTime(double time) {
         this.time = time;
     }
 
-    public int getPrice() {
+    public double getPrice() {
         return price;
     }
 
-    public void setPrice(int price) {
+    public void setPrice(double price) {
         this.price = price;
     }
 }
